+++ src/app/(tools)/tierlistpve/page.tsx
@@ -2,18 +2,21 @@ import type { Metadata } from 'next'
 import TierListPage from './tierlistClient'
 import characters from '@/data/_allCharacters.json' // fichier statique rassemblant tous les persos
 import type { Character } from '@/types/character'
+import { getMonthYear } from '@/utils/getMonthYear';
+
+const monthYear = getMonthYear();
 
 export const metadata: Metadata = {
-  title: 'Outerplane PvE Tier List – May 2025 Update | Outerpedia',
-  description: 'Discover the best characters ranked in Outerplane. Updated for May 2025, curated by the EvaMains Discord community.',
-  keywords: ['Outerplane', 'Tier List', 'Best Characters', 'May 2025', 'EvaMains', 'Outerpedia'],
+  title: `Outerplane PvE Tier List – ${monthYear} Update | Outerpedia`,
+  description: `Discover the best characters ranked in Outerplane. Updated for ${monthYear} , curated by the EvaMains Discord community.`,
+  keywords: ['Outerplane', 'Tier List', 'Best Characters', `${monthYear} `, 'EvaMains', 'Outerpedia'],
   alternates: {
-    canonical: 'https://outerpedia.com/tierlist',
+    canonical: 'https://outerpedia.com/tierlistpve',
   },
   openGraph: {
-    title: 'Outerplane PvE Tier List – May 2025 Update | Outerpedia',
-    description: 'Latest Outerplane character rankings as of May 2025, curated by the EvaMains Discord community.',
-    url: 'https://outerpedia.com/tierlist',
+    title: `Outerplane PvE Tier List – ${monthYear} Update | Outerpedia`,
+    description: `Latest Outerplane character rankings as of ${monthYear} , curated by the EvaMains Discord community.`,
+    url: 'https://outerpedia.com/tierlistpve',
     type: 'website',
     images: [
       {
@@ -26,8 +29,8 @@ export const metadata: Metadata = {
   },
   twitter: {
     card: 'summary',
-    title: 'Outerplane PvE Tier List – May 2025 | Outerpedia',
-    description: 'Best Outerplane characters ranked – May 2025 update curated by EvaMains Discord.',
+    title: `Outerplane PvE Tier List – ${monthYear} | Outerpedia`,
+    description: `Best Outerplane characters ranked – ${monthYear} update curated by EvaMains Discord.`,
     images: ['https://outerpedia.com/images/ui/nav/CM_Lobby_Button_Misson.png'],
   },
 }